feat: 사용자 프로필 파일(사진) 업로드 + 카드 CRD (#13)
* 리뷰 CRUD 구현

* dto와 entity 변환

* 컨트롤러에서 엔티티로 데이터를 받는 것을 dto로 전환

* 리뷰 페이징 기능 추가

* acc

* 숙소 리뷰 CRUD 완성

* 필요없는 엔티티 삭제

* 리뷰 수정 중

* 호텔 JSON에서 호텔에 맞는 리뷰 출력

* 호텔 + 객실 기능

* 예약 기능 추가 시작

* 1

* .

* 객실 예약 기능 구현

* .

* 각 호텔 리뷰 평균

BigDecimal 사용

* 호텔 총 개수 출력

* 리뷰 페이징

* main merge

* db 외래키 수정

* Null 체크

* 사용자 프로필 업로드 + 카드 CRD

* main 병합

+++ src/main/java/com/example/daewoo/review/dto/ReviewEntity.java
@@ -6,6 +6,8 @@ import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -20,7 +22,7 @@ public class ReviewEntity {
 
     private String title;
     private String content;
-    private Integer score;
+    private BigDecimal score;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
